requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
azure-cognitiveservices-speech==1.25.0
Babel==2.9.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
environs==8.0.0
frozenlist==1.3.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.9
libretranslatepy==2.1.1
loader==2017.9.11
lxml==4.9.2
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
openai==0.27.0
packaging==23.0
pydub==0.25.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tqdm==4.64.1
translate==3.6.1
urllib3==1.26.14
utils==1.0.1
yarl==1.8.2
